Modify task 102 code file

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -8,7 +8,7 @@
 int main(void)
 {
 int n;
-long double term0, term1, term;
+long int term0, term1, term;
 term0 = 1;
 term1 = 2;
 n = 0;
